+++ example/entity_store_simple_example/lib/models/todo.dart
@@ -20,21 +20,58 @@ class Todo implements Entity<int> {
   final String name;
   final bool isDone;
 
-  Todo(this.id, this.name, this.isDone);
+  Todo({
+    required this.id,
+    required this.name,
+    required this.isDone,
+  });
 
   factory Todo.create(int id) {
     return Todo(
-      id,
-      getRandomTodoName(),
-      false,
+      id: id,
+      name: getRandomTodoName(),
+      isDone: false,
     );
   }
 
   Todo toggle() {
     return Todo(
-      id,
-      name,
-      isDone ? false : true,
+      id: id,
+      name: name,
+      isDone: isDone ? false : true,
+    );
+  }
+}
+
+class SubTodo implements Entity<int> {
+  @override
+  final int id;
+  final int todoId;
+  final String name;
+  final bool isDone;
+
+  SubTodo({
+    required this.id,
+    required this.name,
+    required this.isDone,
+    required this.todoId,
+  });
+
+  factory SubTodo.create(int id, int todoId) {
+    return SubTodo(
+      id: id,
+      todoId: todoId,
+      name: getRandomTodoName(),
+      isDone: false,
+    );
+  }
+
+  SubTodo toggle() {
+    return SubTodo(
+      id: id,
+      todoId: todoId,
+      name: name,
+      isDone: isDone ? false : true,
     );
   }
 }
@@ -44,13 +81,41 @@ class TodoRepository extends LocalStorageRepository<int, Todo> {
 
   @override
   Todo fromJson(Map<String, dynamic> json) {
-    return Todo(json["id"], json["name"], json["isDone"]);
+    return Todo(
+      id: json["id"],
+      name: json["name"],
+      isDone: json["isDone"],
+    );
   }
 
   @override
   Map<String, dynamic> toJson(Todo entity) {
     return {
       "id": entity.id,
+      "name": entity.name,
+      "isDone": entity.isDone,
+    };
+  }
+}
+
+class SubTodoRepository extends LocalStorageRepository<int, SubTodo> {
+  SubTodoRepository(super.controller, super.localStorageHandler);
+
+  @override
+  SubTodo fromJson(Map<String, dynamic> json) {
+    return SubTodo(
+      id: json["id"],
+      todoId: json["todoId"],
+      name: json["name"],
+      isDone: json["isDone"],
+    );
+  }
+
+  @override
+  Map<String, dynamic> toJson(SubTodo entity) {
+    return {
+      "id": entity.id,
+      "todoId": entity.todoId,
       "name": entity.name,
       "isDone": entity.isDone,
     };