+++ packages/entity_store/lib/src/entity_store/entity_store_provider_scope.dart
@@ -59,4 +59,69 @@ extension BuildContextEntityStoreProviderScope on BuildContext {
   E? readOne<Id, E extends Entity<Id>>(Id id) {
     return read<EntityStoreNotifier>().value.get<Id, E>(id);
   }
+
+  WatcherWhere<E?, void> watchById<Id, E extends Entity<Id>>(Id id) {
+    return _watchById<Id, E, void>(null, (prevData) => id);
+  }
+
+  WatcherWhere<EntityMap<Id, E>, void> watchWhere<Id, E extends Entity<Id>>([
+    bool Function(E?)? test,
+  ]) {
+    return _watchWhere<Id, E, void>(
+      null,
+      (__, entity) => test?.call(entity) ?? true,
+    );
+  }
+
+  WatcherWhere<E, U> _watchById<Id, E extends Entity<Id>, U>(
+    WatcherWhere<U, dynamic>? prevChain,
+    Id Function(U? prevData) idSelector,
+  ) {
+    return WatcherWhere<E, U>._(
+      this,
+      (prevData) => watchOne<Id, E>(idSelector(prevData))!,
+      prevChain?.value,
+    );
+  }
+
+  WatcherWhere<EntityMap<Id, E>, U> _watchWhere<Id, E extends Entity<Id>, U>(
+    WatcherWhere<U, dynamic>? prevChain,
+    bool Function(U? prevData, E entity) test,
+  ) {
+    return WatcherWhere<EntityMap<Id, E>, U>._(
+      this,
+      (prevData) => watchAll<Id, E>(
+        (entity) => test(prevData, entity),
+      ),
+      prevChain?.value,
+    );
+  }
+}
+
+class WatcherWhere<T, U> {
+  final BuildContext _context;
+  final U? _prevData;
+  final T Function(U? prevData) _watch;
+
+  WatcherWhere._(this._context, this._watch, this._prevData);
+
+  T get value => _watch(_prevData);
+
+  WatcherWhere<E, T> watchById<Id, E extends Entity<Id>>(
+    Id Function(T? prevData) idSelector,
+  ) {
+    return _context._watchById(
+      this,
+      (prevData) => idSelector(prevData),
+    );
+  }
+
+  WatcherWhere<EntityMap<Id, E>, T> watchWhere<Id, E extends Entity<Id>>([
+    bool Function(T? prevData, E entity)? test,
+  ]) {
+    return _context._watchWhere<Id, E, T>(
+      this,
+      test ?? (__, ___) => true,
+    );
+  }
 }