+++ packages/entity_store/lib/src/repository_interface/query.dart
@@ -76,13 +76,55 @@ abstract class IRepositoryQuery<Id, E extends Entity<Id>> {
 
   IRepositoryQuery<Id, E> limit(int count);
 
-  IRepositoryQuery<Id, E> startAfterId(Id id);
+  IRepositoryQuery<Id, E> startAfter(Id id);
 
   bool test(Map<String, dynamic> object);
 
-  Future<Result<List<E>, Exception>> findAll();
+  Future<Result<List<E>, Exception>> findAll({
+    FindAllOptions? options,
+  });
 
-  Future<Result<E?, Exception>> findOne();
+  Future<Result<E?, Exception>> findOne({
+    FindOneOptions? options,
+  });
 
   Future<Result<int, Exception>> count();
+
+  List<RepositoryFilter> get filters;
+  List<RepositorySort> get sorts;
+  int? get limitNum;
+  Id? get startAfterId;
+
+  static List<Map<String, dynamic>> findEntities<E extends Entity<Id>, Id>(
+    List<Map<String, dynamic>> objects,
+    IRepositoryQuery<Id, E> query,
+  ) {
+    bool test(Map<String, dynamic> object) {
+      return query.filters.map((e) => e.test(object)).every((e) => e);
+    }
+
+    var entites = objects.where(test);
+
+    if (query.startAfterId != null) {
+      entites = entites.skipWhile((e) => e['id'] != query.startAfterId);
+    }
+
+    if (query.limitNum != null) {
+      entites = entites.take(query.limitNum!);
+    }
+
+    if (query.sorts.isNotEmpty) {
+      final sort = query.sorts.first;
+      entites = entites.toList()
+        ..sort((a, b) {
+          if (sort.descending) {
+            return b[sort.field]!.compareTo(a[sort.field]);
+          } else {
+            return a[sort.field]!.compareTo(b[sort.field]);
+          }
+        });
+    }
+
+    return entites.toList();
+  }
 }